Move login routes to their own test class
